Extract fetchRootData
